+++ core-ng/src/main/java/core/framework/internal/db/QueryImpl.java
@@ -51,12 +51,12 @@ public class QueryImpl<T> implements Query<T> {
     }
 
     @Override
-    public void skip(int skip) {
+    public void skip(Integer skip) {
         this.skip = skip;
     }
 
     @Override
-    public void limit(int limit) {
+    public void limit(Integer limit) {
         this.limit = limit;
     }
 
@@ -65,7 +65,7 @@ public class QueryImpl<T> implements Query<T> {
         if (groupBy != null) throw new Error("fetch must not be used with groupBy, groupBy=" + groupBy);
         if (limit != null && limit == 0) return List.of();  // for pagination search api returns records and count, sometimes it passes limit = 0 to get count only
         String sql = selectQuery.fetchSQL(whereClause, sort, skip, limit);
-        Object[] params = selectQuery.fetchParams(this.params, skip, limit);
+        Object[] params = selectQuery.params(this.params, skip, limit);
         return database.select(sql, entityClass, params);
     }
 
@@ -74,22 +74,21 @@ public class QueryImpl<T> implements Query<T> {
         if (groupBy != null) throw new Error("fetch must not be used with groupBy, groupBy=" + groupBy);
         if (limit != null && limit == 0) return Optional.empty();
         String sql = selectQuery.fetchSQL(whereClause, sort, skip, limit);
-        Object[] params = selectQuery.fetchParams(this.params, skip, limit);
+        Object[] params = selectQuery.params(this.params, skip, limit);
         return database.selectOne(sql, entityClass, params);
     }
 
     @Override
     public <P> List<P> project(String projection, Class<P> viewClass) {
         String sql = selectQuery.projectionSQL(projection, whereClause, groupBy, sort, skip, limit);
-        Object[] params = selectQuery.fetchParams(this.params, skip, limit);
+        Object[] params = selectQuery.params(this.params, skip, limit);
         return database.select(sql, viewClass, params);
     }
 
     @Override
     public <P> Optional<P> projectOne(String projection, Class<P> viewClass) {
-        // project ignores sort, skip and limit, and not report error, mainly for pagination search
-        String sql = selectQuery.projectionSQL(projection, whereClause, groupBy, null, null, null);
-        Object[] params = this.params.toArray();
+        String sql = selectQuery.projectionSQL(projection, whereClause, groupBy, sort, skip, limit);
+        Object[] params = selectQuery.params(this.params, skip, limit);
         return database.selectOne(sql, viewClass, params);
     }
 }